+++ cmd/mosn/main/mosn.go
@@ -57,7 +57,6 @@ import (
 	_ "mosn.io/mosn/pkg/protocol/xprotocol/dubbo"
 	_ "mosn.io/mosn/pkg/protocol/xprotocol/tars"
 	_ "mosn.io/mosn/pkg/router"
-	_ "mosn.io/mosn/pkg/seata"
 	_ "mosn.io/mosn/pkg/stream/http"
 	_ "mosn.io/mosn/pkg/stream/http2"
 	_ "mosn.io/mosn/pkg/stream/xprotocol"
