+++ examples/codes/seata/server_b/service_b_config.json
@@ -53,23 +53,6 @@
 							}
 						}
 					]
-				},
-				{
-					"name":"seataListener",
-					"address": "127.0.0.1:3047",
-					"bind_port": true,
-					"filter_chains": [
-						{
-							"filters": [
-								{
-									"type": "grpc",
-									"config": {
-										"server_name": "seata"
-									}
-								}
-							]
-						}
-					]
 				}
 			]
 		}
